Reworked the meth_parser and meth_lexer to encapsulate the new notion of preconditions as arbitrary Python code. These preconditions are now specified by placing correclty indented Python code between the reserved symbols =BEGIN and =END.
Created action_parser and command_parser, and the module that provides the basic functionality that they encapsulate, fun_map -- this module, given a filename or filehandle or string of python code (all of which must represent a syntactically Python module), dynamically generates a corresponding module, and then creates and returns a map of that module's functions (function_id as str -> function_pointer as Function).

Finally, wrote a PlanningProblem class that provides an abstraction over the process of reading in each of the planning domain file types (which must now be zipped together into an archive), and encapsulates the data these files contain. That is, when initialized with a path to an appropriate zip archive, it comes to contain a representation of an entire RAE/SeRPE planning problem -- with corresponding domain representation, refinement method table, task table, method-task mapping, action model file pointers, and command file pointers.